fix: escHtml on telegram_linked_at title attr; int cast for DateTime modify
Agent-Logs-Url: https://github.com/dsitservicesja-lab/Trading-Web/sessions/013cdd9b-67af-4256-9143-575612af25ec

+++ api/telegram/link_token.php
@@ -75,7 +75,7 @@ try {
     } else {
         /* Generate a new token */
         $token     = bin2hex(random_bytes(32)); // 64 hex chars
-        $expiresAt = (new \DateTime())->modify('+' . (TG_INVITE_EXPIRY_SECONDS / 60) . ' minutes')->format('Y-m-d H:i:s');
+        $expiresAt = (new \DateTime())->modify('+' . (int)(TG_INVITE_EXPIRY_SECONDS / 60) . ' minutes')->format('Y-m-d H:i:s');
 
         $pdo->prepare(
             'INSERT INTO telegram_link_tokens (user_id, token, expires_at) VALUES (?, ?, ?)'
